% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{autocov_gs_cpp}
\alias{autocov_gs_cpp}
\title{Autocovariance function using Sbrana's formula}
\usage{
autocov_gs_cpp(ar_in, ma_in, cov, max_lag)
}
\arguments{
\item{ar_in}{a 3d-array with the p AR coefficient matrices}

\item{ma_in}{a 3d-array with the q MA coefficient matrices}

\item{cov}{covariance matrix of the innovations}

\item{max_lag}{non-negative integer with max lag of the autocovariance function}
}
\value{
A 3d array with the max_lag+1 autocovariance matrices
}
\description{
Autocovariance function using Sbrana's formula
}
