% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{autocov_ly}
\alias{autocov_ly}
\title{Autocovariance function of VARMA process}
\usage{
autocov_ly(varma, maxlag = 10)
}
\arguments{
\item{varma}{a varma object.}

\item{maxlag}{positive integer with the maximum lag for the autocovariance function.}
}
\value{
Array with autocovariance matrices.
}
\description{
It computes the autocovariances using the state-space representation of the VARMA process and
the following formula for the variance of the state variable using the Lyapunov equation solution
of the package netcontrol:
\eqn{vec(\Gamma_0) = (I - T \otimes T)^{-1} vec(RQR')}
}
