% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{autocov_mc}
\alias{autocov_mc}
\title{Autocovariance function of VARMA process}
\usage{
autocov_mc(varma, maxlag = 10)
}
\arguments{
\item{varma}{a varma object.}

\item{maxlag}{positive integer with the maximum lag for the autocovariance function.}
}
\value{
Array with autocovariance matrices.
}
\description{
It computes the autocovariances using Tucker McElroy's algorithm.
The function is based on the original R code made available by the author.
}
