% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{autocov_mi}
\alias{autocov_mi}
\title{Autocovariance function of VARMA process}
\usage{
autocov_mi(varma, maxlag = 10)
}
\arguments{
\item{varma}{a varma object.}

\item{maxlag}{positive integer with the maximum lag for the autocovariance function.}
}
\value{
Array with autocovariance matrices.
}
\description{
It computes the autocovariances using the state-space representation of the VARMA process and
the formula of Mittnik (1990)
}
