% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{booted_irf}
\alias{booted_irf}
\title{IRF from bootstrapped varma objects}
\usage{
booted_irf(
  varma_list,
  maxlag = 10,
  orth = c("none", "cholesky", "spectral", "generalized"),
  title = NULL,
  probs = c(0.1, 0.9),
  varnames = paste0("y", 1:dim(varma_list[[1]])[1]),
  ...
)
}
\arguments{
\item{varma_list}{the list of bootstrapped varma objects
generated by `bootstrap_varma()`.}

\item{maxlag}{the maximum lag of the IRF.}

\item{orth}{orthogonalization method.}

\item{title}{title for the plot (default: NULL, that is, no title).}

\item{probs}{a vector with two values between 0 and 1 representing.
the percentiles to be represented by the fan in the plot.}

\item{varnames}{names of the variables (default: y1, y2, ...).}

\item{...}{other parameters to be passed to the `irf()` function.}
}
\value{
Nothing: it produces plots.
}
\description{
It produces IRF plots from bootstrapped varma
estimates produced by `bootstrap_varma()`.
}
