% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bootstrap_varma}
\alias{bootstrap_varma}
\title{Simulate from an estimated varma by bootstrapping residuals}
\usage{
bootstrap_varma(
  varma,
  nsim = 100,
  n = varma$n,
  resids_to_skip = 5,
  fit_fn = get(varma$estimation_method),
  ...,
  parallel = FALSE,
  n_cores = NULL
)
}
\arguments{
\item{varma}{a varma object containing residuals.}

\item{nsim}{number of sample paths to simulate (default: 100).}

\item{n}{length of the time series (default: the `n` in the varma object).}

\item{resids_to_skip}{non-negative integer witht he number of initial residuals to drop because
they may have a very large variance (default: 5).}

\item{fit_fn}{function to apply to every simulated path: it is usually a
`fit_varma_???` function (default: the function that generated the estimates
in the varma object).}

\item{...}{other parameters to be passed to `fit_fn`.}

\item{parallel}{logical, if TRUE, parallel processing is used (default: FALSE).}

\item{n_cores}{number of cores for parallel processing. If NULL, uses
parallel::detectCores() - 1 (default: NULL).}
}
\value{
A list with `nsim` slots containing the applications of the
function `fit_fn` to the simulated sample paths.
}
\description{
It produces sample paths of a varma process by bootstrapping the residuals
of an estimated varma. A function (usually a fit_varma function) is applied
to the simulated sample paths.
}
