% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{companion_form}
\alias{companion_form}
\title{Companion form of a VARMA}
\usage{
companion_form(varma, part = c("ar", "ma"))
}
\arguments{
\item{varma}{a varma object.}

\item{part}{vector of strings: can be `"ar"`, `"ma"` or both `c("ar", "ma")`,
indicating which part should ne cast in companion form.}
}
\description{
It computes the companion form of the AR and of the MA part.
}
