% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{convert_fable_varma}
\alias{convert_fable_varma}
\title{Cast VARMA estimates from fable::VARIMA into a varma object}
\usage{
convert_fable_varma(object)
}
\arguments{
\item{object}{object returned by a code like
`model(Y, VARIMA(vars(V1, V2, V3) ~ pdq(1, 0, 2), identification = "kronecker_indices"))`.}
}
\value{
A varma object.
}
\description{
It takes the output of the function VARMA of the fable package
and turns it into a varma object.
}
