% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{convert_mts_varma}
\alias{convert_mts_varma}
\title{Cast VARMA estimates from MTS::VARMA into a varma object}
\usage{
convert_mts_varma(mts_varma)
}
\arguments{
\item{mts_varma}{list returned by MTS::VARMA.}
}
\value{
A varma object.
}
\description{
It takes the output of the function VARMA of the MTS package
and turns it into a varma object.
}
