% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dim.varma}
\alias{dim.varma}
\title{Dimensions of a varma object}
\usage{
\method{dim}{varma}(x)
}
\arguments{
\item{x}{a varma object.}
}
\value{
A vector of length 3 with the dimensions of the varma object:
number of time seires (m), order of AR part (p), order of MA part (q).
}
\description{
Dimensions of a varma object
}
