% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimate_varma_gls_cpp}
\alias{estimate_varma_gls_cpp}
\title{GLS estimation of VARMA using the diagonal MA or final MA form}
\usage{
estimate_varma_gls_cpp(Y, U, SigmaU_inv, p, q, type)
}
\arguments{
\item{Y}{data matrix}

\item{U}{innovation matrix}

\item{SigmaU_inv}{inverse of the innovation covariance matrix}

\item{p}{order of VAR part}

\item{q}{order of VMA part}

\item{type}{a string with one of the two choices "diagonal", "final"}
}
\description{
GLS estimation of VARMA using the diagonal MA or final MA form
}
