% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_varma_dpf}
\alias{fit_varma_dpf}
\title{Fit a VARMA(p,q) in final MA form using the iterated GLS method}
\usage{
fit_varma_dpf(
  Y,
  p = 1,
  q = 1,
  intercept = TRUE,
  r = max(p + q, round(nrow(Y)/(4 * ncol(Y)))),
  verbose = FALSE
)
}
\arguments{
\item{Y}{a numeric matrix or ts object with d columns (series) and N rows (observations).}

\item{p}{the autoregressive order.}

\item{q}{the moving average order.}

\item{intercept}{logical: if TRUE (default) a vector of intercepts is computed.}

\item{r}{order of the long VAR(r) for approximating the residuals at first step}

\item{verbose}{logical, if TRUE some textual output is printed during execution}
}
\value{
A varma object.
}
\description{
Estimates the parameters of a VARMA(p,q) model by maximizing the
likelihood iterating GLS step as in Dufour and Pelletier (2022).
}
