% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_varma_fkf}
\alias{fit_varma_fkf}
\title{VARMA estimation using Maximum-Likelihoo in State-Space form (FKF.SP)}
\usage{
fit_varma_fkf(
  Y,
  p = 1,
  q = 1,
  intercept = TRUE,
  maxit = 100,
  ret = c("varma", "ss")
)
}
\arguments{
\item{Y}{a (n x m) matrix with the time series.}

\item{p}{order of the AR part.}

\item{q}{order of the MA part.}

\item{intercept}{boolean, if TRUE, the model will include an intercept.}

\item{maxit}{integer, maximum number of iterations for the iterated Hannan-Rissanen
estimator, which is used for providing starting values for numerical MLE and for
computing the marginal covariance matrix of the VARMA in state-space form, used
for initializing the Kalman filter.}

\item{ret}{a character string indicating the type of output: "varma" (default)
or "ss" (a list with the state-space form matrices and fkf function output).}
}
\value{
A varma object or a list with the matrices of the state-space form,
according to the choice of the `ret` parameter.
}
\description{
VARMA estimation using Maximum-Likelihoo in State-Space form (FKF.SP)
}
