% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_varma_ihr}
\alias{fit_varma_ihr}
\title{VARMA estimation using the iterated Hannan-Rissanen method}
\usage{
fit_varma_ihr(
  Y,
  p = 1,
  q = 1,
  intercept = TRUE,
  maxdiff = 1e-05,
  maxit = 100,
  r = max(p + q, round(nrow(Y)/(4 * ncol(Y)))),
  ret = c("varma", "regression")
)
}
\arguments{
\item{Y}{a (n x m) matrix with the time series.}

\item{p}{order of the AR part.}

\item{q}{order of the MA part.}

\item{intercept}{logical, if TRUE, the model includes an intercept.}

\item{maxdiff}{maximum difference between the coefficients of two consecutive iterations:
it is used as a threshold to stop the iterations.}

\item{maxit}{maximum number of iterations.}

\item{r}{integer order of the first step VAR(r) model.}

\item{ret}{a character string indicating the type of output: "varma" or "regression".}
}
\value{
If ret = "varma" a varma object, otherwise a list with the following elements:
coefficients, matrix of residuals, fitted.values, effects, weights, rank,
df.residual, qr, n_iter, res_cov.
}
\description{
VARMA estimation using the iterated Hannan-Rissanen method
}
