% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_varma_kfas}
\alias{fit_varma_kfas}
\title{VARMA estimation using Maximum-Likelihood in State-Space form using KFAS}
\usage{
fit_varma_kfas(
  Y,
  p = 1,
  q = 1,
  intercept = TRUE,
  maxit = 100,
  ret = c("varma", "kfas")
)
}
\arguments{
\item{Y}{a (n x m) matrix with the time series.}

\item{p}{order of the AR part.}

\item{q}{order of the MA part.}

\item{intercept}{logical, if TRUE, the model will include an intercept.}

\item{maxit}{integer, maximum number of iterations for the iterated Hannan-Rissanen
estimator, which is used for providing starting values for numerical MLE and for
computing the marginal covariance matrix of the VARMA in state-space form, used
for initializing the Kalman filter.}

\item{ret}{a character string indicating the type of output: "varma" or
"kfas" (a KFAS' SSModel object).}
}
\value{
If ret = "varma" a varma object, otherwise a SSModel object from the KFAS package.
}
\description{
VARMA estimation using Maximum-Likelihood in State-Space form using KFAS
}
