% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_varma_rby}
\alias{fit_varma_rby}
\title{Fit a VARMA(p,q) Model using the iterated GLS method}
\usage{
fit_varma_rby(Y, p = 1, q = 1, intercept = TRUE, maxit = 500, tol = 1e-05)
}
\arguments{
\item{Y}{a numeric matrix or ts object with d columns (series) and N rows (observations).}

\item{p}{the autoregressive order.}

\item{q}{the moving average order.}

\item{intercept}{logical: if TRUE (default) a vector of intercepts is computed.}

\item{maxit}{integer, maximum number of GLS iterations.}

\item{tol}{tolerance to determine convergence.}
}
\value{
A varma object.
}
\description{
Estimates the parameters of a VARMA(p,q) model by maximizing the
likelihood iterating GLS step as in Reinsel, Basu and Yap (1992).
}
