% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_varma_skf}
\alias{fit_varma_skf}
\title{Fit a VARMA(p,q) Model using Sparse Kalman Filter}
\usage{
fit_varma_skf(Y, p, q, intercept = TRUE, maxit = 100)
}
\arguments{
\item{Y}{a numeric matrix or ts object with d columns (series) and N rows (observations).}

\item{p}{the autoregressive order.}

\item{q}{the moving average order.}

\item{intercept}{logical: if TRUE (default) a vector of intercepts is computed.}

\item{maxit}{integer, maximum number of iterations for the iterated Hannan-Rissanen
estimator, which is used for providing starting values for numerical MLE and for
computing the marginal covariance matrix of the VARMA in state-space form, used
for initializing the Kalman filter}
}
\value{
A varma object.
}
\description{
Estimates the parameters of a VARMA(p,q) model by maximizing the
likelihood computed with an efficient C++/Armadillo-based Kalman filter.
}
