% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{inv_roots}
\alias{inv_roots}
\title{Inverse roots of the characteristic polynomials of VARMA process}
\usage{
inv_roots(varma, part = c("ar", "ma"), plot = FALSE)
}
\arguments{
\item{varma}{a varma object.}

\item{part}{a character vector indicating which polynomial equation has
to be solved: it can be `"ar"`, `"ma"` or both `c("ar", "ma")`.}

\item{plot}{logical, if TRUE, the function will plot the roots in the complex plane.}
}
\value{
A list with the AR roots (`$ar`) and the MA roots (`$ma`).
}
\description{
It computes the inverse roots of the AR and MA parts and produces
plots of the roots in the complex plane, together with the unit circle.
}
