% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{irf}
\alias{irf}
\title{Impulse response function of VARMA process}
\usage{
irf(
  varma,
  maxlag = 10,
  orth = c("none", "cholesky", "spectral", "generalized"),
  plot = FALSE,
  title = NULL,
  varnames = paste0("y", 1:dim(varma)[1])
)
}
\arguments{
\item{varma}{a varma object.}

\item{maxlag}{positive integer indicating the maximum lag for which the IRF
is computed.
`lagmax`.}

\item{orth}{a character string indicating the orthogonalization method to be used:
"none", "cholesky", "spectral", "generalized" are the available options.}

\item{plot}{logical, if TRUE, the function plots the IRFs.}

\item{title}{string with a title for the plot (default to NULL, that is, no title).}

\item{varnames}{vector of strings with the names of the variables.}
}
\value{
A 3D array with the impulse response functions of the VARMA process.
}
\description{
Impulse response function of VARMA process
}
