% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is_identified}
\alias{is_identified}
\title{Check if the VARMA process is identified}
\usage{
is_identified(varma, tol = 1e-05)
}
\arguments{
\item{varma}{a varma object.}

\item{tol}{a small numeric value indicating the tolerance for the numerical check
(default: 1e-5).}
}
\value{
TRUE if the VARMA is identified and FALSE otherwise.
}
\description{
The function checks if the VARMA process is identified by numerically checking if the
roots of the characteristic polynomials are distinct.
}
