% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{kronecker_indices}
\alias{kronecker_indices}
\title{Computes the Kronecker indices of a VARMA model}
\usage{
kronecker_indices(varma, tol = 1e-08)
}
\arguments{
\item{varma}{an object of varma class (only the slots ar and ma are used)}

\item{tol}{relative singular-value threshold for numerical rank (default \code{1e-8})}
}
\value{
A vector with number of elements equal to the number of time series
with the Kronecker indices.
}
\description{
Computes the Kronecker indices of a VARMA model
}
