% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{logLik.varma}
\alias{logLik.varma}
\title{logLik method for a varma object}
\usage{
\method{logLik}{varma}(object, ...)
}
\arguments{
\item{object}{a varma object.}

\item{...}{not used.}
}
\value{
A numeric scalar with the log-likelihood value and the
attribute df with the model's degrees of freedom.
}
\description{
logLik method for a varma object
}
