% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nobs.varma}
\alias{nobs.varma}
\title{nobs method for a varma object}
\usage{
\method{nobs}{varma}(object, ...)
}
\arguments{
\item{object}{a varma object.}

\item{...}{not used.}
}
\value{
A numeric scalar with the number of observations.
}
\description{
nobs method for a varma object
}
