% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict.varma}
\alias{predict.varma}
\title{predict method for a varma object}
\usage{
\method{predict}{varma}(object, n.ahead = 1, cov = TRUE, ...)
}
\arguments{
\item{object}{a varma object with the estimated model.}

\item{n.ahead}{integer forecast horizon.}

\item{cov}{logical, if TRUE the covariance matrices of the forecasts are produced.}

\item{...}{not used.}
}
\value{
If `cov = FALSE`, a matrix with the forecasts, if `cov = TRUE`
a list with the slot `mean`, containing a matrix with the point forecasts and
the slot `cov` with a 3D array with the cvariance matrices of the forecasts.
}
\description{
The varma object has to be estimated using some of the methods
in this package.
}
