% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.varma}
\alias{print.varma}
\title{Print method for varma object}
\usage{
\method{print}{varma}(x, ...)
}
\arguments{
\item{x}{a varma object.}

\item{...}{not used.}
}
\value{
Nothing.
}
\description{
Print method for varma object
}
