% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{residuals.varma}
\alias{residuals.varma}
\title{residuals method for varma object}
\usage{
\method{residuals}{varma}(object, Y = NULL, recompute = TRUE, ...)
}
\arguments{
\item{object}{a varma object.}

\item{Y}{data matrix if NULL the y slot in the varma object is used}

\item{recompute}{logical, if FALSE and the varma object contains residuals,
the function returns those residuals, if TRUE the function recomputes the
residuals.}

\item{...}{not used.}
}
\value{
Matrix of residuals. If recompute is TRUE, the residuals are always
recomputed by this function, if it is FALSE and the varma object slot
residuals is not NULL, then these residuals are returned.
}
\description{
residuals method for varma object
}
