% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{select_varma_order}
\alias{select_varma_order}
\title{Selct the VARMA order}
\usage{
select_varma_order(
  Y,
  p_max = 5,
  q_max = 5,
  intercept = TRUE,
  ic = c("AIC", "AICc", "BIC", "HQC"),
  fit_fn = fit_varma_ihr,
  ...
)
}
\arguments{
\item{Y}{matrix of time series.}

\item{p_max}{maximum order for the VAR part.}

\item{q_max}{maximum order for the VMA part.}

\item{intercept}{if FALSE assume no intercept is needed, otherwise.
computes the information criterion both with and without intercept.}

\item{ic}{information criterion to use among "AIC", "AICc", "BIC", "HQC".}

\item{fit_fn}{fit_varma_??? function to use.}

\item{...}{parameters passed to the fit_varma_??? function}
}
\value{
A data.frame with columns p, q, intercept, ic
}
\description{
It computes an information criterion for all combinations of
p, q, and intercept for determining the VARMA order.
}
