% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sim_varma}
\alias{sim_varma}
\title{Simulate VARMA process}
\usage{
sim_varma(varma, n = 200, burn_in = 100, var_in_rows = FALSE)
}
\arguments{
\item{varma}{a varma object.}

\item{n}{number of observations to be simulated.}

\item{burn_in}{number of initial observations to discard.}

\item{var_in_rows}{boolean, if TRUE, the variables are in rows,
otherwise they are in columns (the internal Rcpp function works with
variables in rows, so TRUE makes the function slightly faster).}
}
\value{
A matrix with the simulated data.
}
\description{
Simulate VARMA process
}
