% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{varma}
\alias{varma}
\title{varma object class}
\usage{
varma(
  intercept = NULL,
  mean = NULL,
  ar = NULL,
  ma = NULL,
  cov = NULL,
  chol_cov = NULL,
  estimation_method = NULL,
  loglik = NULL,
  n = NULL,
  nobs = NULL,
  npar = NULL,
  y = NULL,
  residuals = NULL,
  state_pred_mean = NULL,
  state_pred_cov = NULL,
  transition_matrix = NULL,
  disturbance_cov = NULL
)
}
\arguments{
\item{intercept}{\eqn{m}-vector with the intercepts.}

\item{mean}{\eqn{m}-vector with means (typically in alternative to `intercept`).}

\item{ar}{array of dimensions \eqn{m \times m \times p} containing the VAR matrices.}

\item{ma}{array of dimensions \eqn{m \times m \times q} containing the VMA matrices.}

\item{cov}{\eqn{m\times m} covariance matrix of the innovations.}

\item{chol_cov}{Cholesky factor of the covariance matrix of the innovations: this is rearly used,
but if some function working with varma objects produces it, it is good to store it to
make its reusable by other functions, that can be more efficient.}

\item{estimation_method}{string with the name of the function that produced the estimates.}

\item{loglik}{value of the log-likelihood function for the estimated VARMA.}

\item{n}{number of time-points of the time series (regardless of possible missing observations).}

\item{nobs}{total number of observations (if no missing observations are present, `nobs`\eqn{=nm}).}

\item{npar}{number of estimated parameters.}

\item{y}{\eqn{n\times m} or \eqn{n-p\times m} data matrix.}

\item{residuals}{\eqn{n\times m} or \eqn{n-p\times m} residuals matrix.}

\item{state_pred_mean}{mean of the one-step-ahead predicted state vector.
at time \eqn{t = n+1} (relevant if the VARMA was estimated in state-space form).}

\item{state_pred_cov}{covariance matrix of the one-step-ahead predicted state vector
at time \eqn{t = n+1} (relevant if the VARMA was estimated in state-space form).}

\item{transition_matrix}{transition matrix of the VARMA in state space
(relevant if the VARMA was estimated in state-space form).}

\item{disturbance_cov}{covariance matrix of the state disturbance vector
(relevant if the VARMA was estimated in state-space form).}
}
\description{
This constructor function has been created just to illustrate the structure
of the varma object class. The varma object class is designed to contain
all the information needed to use VARMA models at various levels.
You can just use it to store a VARMA process definition or to contain
an estimated VARMA model with data, goodness of fit, residuals, etc.
A VARMA process definition can be used to produce simulated sample paths,
population quantities (IRFs, characeteristic roots, identification, etc.).
An estimated VARMA model can be used to produce the sample counterparts of the
population quantities, bootstapped sample paths, information criteria, and forecasts.
In the following we use the following symbols: \eqn{m}, \eqn{p}, \eqn{q}, \eqn{n} represent
the dimension of the time series vector, the largest lag of the VAR part of the model,
the largest lag of the VMA part of the mode, and the number of time series points, respectively.
}
