% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{varma_to_ss}
\alias{varma_to_ss}
\title{VARMA in state space form}
\usage{
varma_to_ss(varma, form = c("harvey", "pearlman"))
}
\arguments{
\item{varma}{varma object.}

\item{form}{string with the name of the state-space form to use: "harvey" or "pearlman".}
}
\value{
A list with the following matrices of the state-space form:
T, R, Q, Z.
}
\description{
It casts the VARMA model in state space form:
}
\details{
\eqn{y_t = Z_t \alpha_t}
\eqn{\alpha_{t+1} = T \alpha_t + R \eta_t}
where \eqn{\eta_t} is a white noise process with covariance matrix \eqn{Q}.
}
