% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_varma_dma}
\alias{fit_varma_dma}
\title{VARMA estimation in Diagonal MA form using the method of Dufour and Pelletier}
\usage{
fit_varma_dma(
  Y,
  p = 1,
  q = 1,
  intercept = TRUE,
  r = max(p + q, round(nrow(Y)/(4 * ncol(Y)))),
  ret = c("varma", "regression")
)
}
\arguments{
\item{Y}{a (n x m) matrix with the time series.}

\item{p}{order of the AR part.}

\item{q}{order of the MA part.}

\item{intercept}{logical, if TRUE, the model includes an intercept.}

\item{r}{integer order of the first step VAR(r) model.}

\item{ret}{a character string indicating the type of output: "varma" or "regression".}
}
\value{
If ret = "varma" a varma object, otherwise a list with the following elements:
coefficients, matrix of residuals, fitted.values, effects, weights, rank,
df.residual, qr, n_iter, res_cov.
}
\description{
VARMA estimation in Diagonal MA form using the method of Dufour and Pelletier
}
