% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_varma_net}
\alias{fit_varma_net}
\title{VARMA estimation using the iterated Hannan-Rissanen method with regularization}
\usage{
fit_varma_net(
  Y,
  p = 1,
  q = 1,
  intercept = TRUE,
  lambda = NULL,
  alpha = 1,
  relax = TRUE,
  gamma = 0,
  sur = TRUE,
  maxdiff = 1e-05,
  maxit = 100,
  r = max(p + q, round(nrow(Y)/(4 * ncol(Y)))),
  parallel = FALSE,
  n_cores = NULL,
  ret = c("varma", "regression")
)
}
\arguments{
\item{Y}{a (n x m) matrix with the time series.}

\item{p}{order of the AR part.}

\item{q}{order of the MA part.}

\item{intercept}{boolean, if TRUE, the model will include an intercept.}

\item{lambda}{positive numeric regularization parameter.}

\item{alpha}{value in the interval [0, 1]; 1 for pure LASSO and 0 for pure RIDGE.}

\item{relax}{logical, if FALSE the classical elastic net is used, if TRUE (default)
the estimations and cross validation are done also using regression coefficients
computed as \eqn{(1-\gamma)\beta^{OLS}_i + \gamma \beta_i}, where \eqn{i} ranges
over the set of non-zero coefficients and \eqn{\beta^{OLS}_i} denotes the
regression coefficient computed by OLS (using only the active regressors).}

\item{gamma}{vector of weights in the range \eqn{[0, 1]} to be used when.
`relax = TRUE`. The default is 0, corresponding to Post-LASSO OLS.}

\item{sur}{logical, if TRUE a final seemingly unrelated regressions
step is computed, otherwise the equation by equation estimate is provided.}

\item{maxdiff}{maximum difference between the coefficients of two consecutive
iterations in the first step based on `fit_varma_ihr`:
it is used as a threshold to stop the iterations.}

\item{maxit}{maximum number of iterations to be used in the first step
based on `fit_varma_ihr`.}

\item{r}{integer order of the first step VAR(r) model used in `fit_varma_ihr`.}

\item{ret}{a character string indicating the type of output: "varma" or "regression".}
}
\value{
If ret = "varma" a varma object, otherwise a list with the results of
cv.glmnet for each response variable.
}
\description{
VARMA estimation using the iterated Hannan-Rissanen method with regularization
}
