% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{irf_distance}
\alias{irf_distance}
\title{Compute the distance between two IRFs}
\usage{
irf_distance(irf1, irf2, r = 2, omit_lag0 = TRUE)
}
\arguments{
\item{irf1}{an array of dimensions m x m x max_lags containing an IRF.}

\item{irf2}{an array of dimensions m x m x max_lags containing an IRF.}

\item{r}{positive number, order of the norm to be used (default is 2 -> Euclidean norm).}

\item{omit_lag0}{logical, do not consider the IRFs at lag-0 (default TRUE).}
}
\value{
Distance between two IRFs (a positive number).
}
\description{
It uses the Lp norm to compute the distance between
two IRFs (the order-0 IRF can be omitted)
}
