% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rvarma}
\alias{rvarma}
\title{Generate Coefficients for a Stationary and Invertible VARMA(p, q) Process}
\usage{
rvarma(
  m,
  p,
  q,
  max_eig_ar = 0.9,
  max_eig_ma = 0.9,
  dist = function(n) runif(n, -1, 1)
)
}
\arguments{
\item{m}{integer with the dimension of the time series process (m > 0).}

\item{p}{integer with the order of the autoregressive (AR) part (p >= 0).}

\item{q}{integer with the order of the moving average (MA) part (q >= 0).}

\item{max_eig_ar}{numeric value between 0 and 1 indicating the desired maximum
eigenvalue modulus for the AR companion matrix to ensure stationarity.}

\item{max_eig_ma}{numeric value between 0 and 1 with the desired maximum
eigenvalue modulus for the MA companion matrix to ensure invertibility.}

\item{dist}{function to generate the random coefficients.}
}
\value{
A list with two components:
\describe{
  \item{\code{ar}}{An m x m x p 3-dimensional array of AR coefficient matrices.}
  \item{\code{ma}}{An m x m x q 3-dimensional array of MA coefficient matrices.}
}
}
\description{
This function generates random coefficient matrices for a m-dimensional
Vector Autoregressive Moving Average (VARMA) process of specified orders p and q.
The user can control the stability of the process by setting the maximum
modulus of the eigenvalues of the AR and MA companion matrices, ensuring
stationarity and invertibility respectively.
}
\details{
The function ensures stationarity (for the AR part) and invertibility (for
the MA part) by generating random coefficient matrices, constructing the
corresponding companion matrix, and then rescaling the coefficients. The
rescaling is done such that the largest eigenvalue modulus of the companion
matrix matches the user-specified value (`max_eig_ar` or `max_eig_ma`).
The method relies on the property that scaling the i-th coefficient matrix (A_i)
by s^i results in scaling the eigenvalues of the companion matrix by s.
}
\examples{
mod <- rvarma(m = 3, p = 2, q = 2, max_eig_ar = 0.95, max_eig_ma = 0.95)

}
